Simple changes to default app that show suspend fail.
With this code a toast is shown every 2 seconds. When the suspend button in the debugger is pressed the toasts to not stop.

When the program is run normally in Windows, the toasts stop as soon as the app is suspended. The app is suspended by minimizing the window and waiting 10 seconds.